refactor(ui): hoist PageContainer into @unkey/ui as a compound component
Move the PageContainer wrapper out of the dashboard and into @unkey/ui
alongside PageHeader and PageBody. Replace its header prop with a
composition API: width="default|full" on PageContainer flows to a
data-width attribute and group/page class, and PageHeader/PageBody style
themselves with group-data variants, mirroring the Sidebar primitive.

Update the four settings pages to compose the parts, and rebuild the
design docs layout examples as standalone files rendered through Preview,
which gains a bleed variant for full-page layouts.
