+++ web/apps/dashboard/app/(app)/[workspaceSlug]/settings/team/page.tsx
@@ -1,7 +1,6 @@
-import { PageContainer } from "@/components/page-header/page-container";
 import { getAuth } from "@/lib/auth";
 import { db } from "@/lib/db";
-import { PageHeader, PageHeaderContent, PageHeaderTitle } from "@unkey/ui";
+import { PageBody, PageContainer, PageHeader, PageHeaderContent, PageHeaderTitle } from "@unkey/ui";
 import { TeamPageClient } from "./client";
 
 export const revalidate = 0;
@@ -15,23 +14,22 @@ export default async function SettingTeamPage() {
 
   const team = workspace?.quotas?.team ?? false;
 
-  const header = (
-    <PageHeader>
-      <PageHeaderContent>
-        <PageHeaderTitle>Team</PageHeaderTitle>
-      </PageHeaderContent>
-    </PageHeader>
-  );
-
   return (
-    <PageContainer header={header}>
-      {workspace ? (
-        <div className="w-full flex flex-col pt-4">
-          <TeamPageClient team={team} />
-        </div>
-      ) : (
-        <div>Workspace not found</div>
-      )}
+    <PageContainer>
+      <PageHeader>
+        <PageHeaderContent>
+          <PageHeaderTitle>Team</PageHeaderTitle>
+        </PageHeaderContent>
+      </PageHeader>
+      <PageBody>
+        {workspace ? (
+          <div className="w-full flex flex-col pt-4">
+            <TeamPageClient team={team} />
+          </div>
+        ) : (
+          <div>Workspace not found</div>
+        )}
+      </PageBody>
     </PageContainer>
   );
 }
