+++ web/internal/ui/src/components/page-header/index.ts
@@ -1,3 +1,4 @@
-export * from "./page-header";
 export * from "./page-body";
+export * from "./page-container";
+export * from "./page-header";
 export * from "./secondary-nav";
