+++ web/internal/ui/src/components/page-header/page-header.tsx
@@ -3,7 +3,15 @@ import { cn } from "../../lib/utils";
 
 const PageHeader = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
   ({ className, ...props }, ref) => (
-    <div ref={ref} className={cn("flex items-start justify-between gap-3", className)} {...props} />
+    <div
+      ref={ref}
+      className={cn(
+        "mx-auto flex w-full max-w-7xl items-start justify-between gap-3 px-4 pt-6",
+        "group-data-[width=full]/page:max-w-none group-data-[width=full]/page:border-b group-data-[width=full]/page:border-grayA-4 group-data-[width=full]/page:pb-4 group-data-[width=full]/page:pt-4",
+        className,
+      )}
+      {...props}
+    />
   ),
 );
 PageHeader.displayName = "PageHeader";
